pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hie"
version = "0.1.0"
description = "Healthcare Integration Engine - Enterprise-grade healthcare messaging platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "HIE Contributors" }
]
keywords = ["healthcare", "integration", "hl7", "fhir", "messaging", "nhs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "aiohttp>=3.9",
    "asyncpg>=0.29",
    "redis>=5.0",
    "msgpack>=1.0",
    "structlog>=24.0",
    "click>=8.1",
    "watchfiles>=0.21",
    "prometheus-client>=0.19",
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
    "black>=24.0",
    "pre-commit>=3.6",
    "httpx>=0.26",
]

hl7 = [
    "hl7>=0.4.5",
]

fhir = [
    "fhir.resources>=7.0",
]

all = [
    "hie[dev,hl7,fhir]",
]

[project.scripts]
hie = "hie.cli:main"

[project.urls]
Homepage = "https://github.com/hie/hie"
Documentation = "https://hie.readthedocs.io"
Repository = "https://github.com/hie/hie"

[tool.hatch.build.targets.wheel]
packages = ["hie"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["hie"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --cov=hie --cov-report=term-missing"

[tool.coverage.run]
source = ["hie"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
